Fix the parsing functions: incomplete parses should not be accepted
This fixes an issue when parsing maxsat output: Yices prefixes the model output with a line specifying how many maxsat goals were not satisfied.
This confuses our parser which only parses this line (as garbage) and then gives up, ignoring the actual model specification.
With this change the parser will refuse to continue and error out, instead of returning a wrong value.