Filter maxsat info lines from the output.
I don't like this solution, it is flimsy. Yices could send all kinds of garbage to std output, and all this patch does is filter one instance of that.
I like better the approach of parsing line by line and then throwing away all the failed parses, but we cannot assume that every item of the model is going to fit in a single line.